backend/editgenerator/pyproject.toml
[project]
name = "editgenerator"
version = "0.1.0"
description = "AI-powered video edit generator"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.21.0",
    "python-dotenv>=1.0.0",
    "jsonschema>=4.20.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0"
]
